Add: ajv decorators
