+++ package.json
@@ -63,6 +63,7 @@
     "dotenv": "10.0.0",
     "fastify": "^3.25.0",
     "fastify-swagger": "^4.12.6",
+    "nestjs-ajv-glue": "^1.0.4",
     "nodemon": "^2.0.15",
     "pg": "^8.7.1",
     "pino": "^7.6.1",
