+++ src/boards/validators/board-select-fields.ts
@@ -1,0 +1,18 @@
+export const boardSelectFields = {
+  type: 'object',
+  properties: {
+    id: { type: 'string' },
+    title: { type: 'string' },
+    columns: {
+      type: 'array',
+      items: {
+        type: 'object',
+        properties: {
+          id: { type: 'string' },
+          title: { type: 'string' },
+          order: { type: 'number' },
+        },
+      },
+    },
+  },
+};
