+++ src/main.ts
@@ -3,6 +3,7 @@ import {
   FastifyAdapter,
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
+import { AjvValidationPipe } from 'nestjs-ajv-glue';
 import { AppModule } from './app.module';
 import { config } from './common/config';
 
@@ -11,6 +12,10 @@ async function bootstrap() {
     AppModule,
     new FastifyAdapter({ logger: true })
   );
+
+  /** Добавить @AjvQuery @AjvBody @AjvParams валидаторы */
+  app.useGlobalPipes(new AjvValidationPipe());
+
   await app.listen(config.PORT, '0.0.0.0');
 }
 bootstrap();
