--- src_old/resources/boards/api/schema/get-by-id.schema.ts
@@ -1,14 +1,0 @@
-import { uuidAsParam } from '../../../common/validators/uuid-as-param';
-import { boardSelectFields } from '../../validators/board-select-fields';
-
-export const getByIdSchema = {
-  schema: {
-    ...uuidAsParam,
-    response: {
-      200: {
-        type: 'object',
-        properties: boardSelectFields,
-      },
-    },
-  },
-};
