--- src_old/resources/boards/api/schema/get.schema.ts
@@ -1,15 +1,0 @@
-import { boardSelectFields } from '../../validators/board-select-fields';
-
-export const getSchema = {
-  schema: {
-    response: {
-      200: {
-        type: 'array',
-        items: {
-          type: 'object',
-          properties: boardSelectFields,
-        },
-      },
-    },
-  },
-};
