--- src_old/resources/boards/api/schema/put.schema.ts
@@ -1,18 +1,0 @@
-import { uuidAsParam } from '../../../common/validators/uuid-as-param';
-import { boardInsertFields } from '../../validators/board-insert-fields';
-import { boardSelectFields } from '../../validators/board-select-fields';
-
-export const putSchema = {
-  schema: {
-    ...uuidAsParam,
-    body: {
-      ...boardInsertFields,
-    },
-    response: {
-      200: {
-        type: 'object',
-        properties: boardSelectFields,
-      },
-    },
-  },
-};
