+++ src_old/resources/boards/board.router.ts
@@ -4,10 +4,10 @@ import { get } from './api/get';
 import { getById } from './api/get-by-id';
 import { post } from './api/post';
 import { put } from './api/put';
-import { getByIdSchema } from './api/schema/get-by-id.schema';
-import { getSchema } from './api/schema/get.schema';
-import { postSchema } from './api/schema/post.schema';
-import { putSchema } from './api/schema/put.schema';
+import { getByIdSchema } from '../../../src/boards/schema/get-by-id.schema';
+import { getSchema } from '../../../src/boards/schema/get.schema';
+import { postSchema } from '../../../src/boards/schema/post.schema';
+import { putSchema } from '../../../src/boards/schema/put.schema';
 
 export const boardRouter = async (fastify: FastifyInstance) => {
   fastify.get('', getSchema, get);
