--- src_old/resources/boards/validators/board-select-fields.ts
@@ -1,15 +1,0 @@
-export const boardSelectFields = {
-  id: { type: 'string' },
-  title: { type: 'string' },
-  columns: {
-    type: 'array',
-    items: {
-      type: 'object',
-      properties: {
-        id: { type: 'string' },
-        title: { type: 'string' },
-        order: { type: 'number' },
-      },
-    },
-  },
-};
